C# 2D Section 4 (#5)
* Initialize project

* Complete Lecture 51

Added the previsualization canvas and the background canvas.

* Complete Lecture 52

Added TextMeshPro and font assets.

* Complete Lecture 53

Added buttons to the quiz canvas.

* Complete Lecture 54

Added the Question scriptable object and several question assets.

* Complete Lectures 55 and 56

Added answers to the Question scriptable object and filled in the answer values of the question assets.

* Complete Lecture 57

Added the QuizController script to fill in question text from the scriptable objects.

* Complete Lecture 58

Added a loop to fill in all the button texts.

* Complete Lecture 59

Added feature to highlight correct and wrong answers.

* Complete Lecture 60

Added functions to modify the answer buttons' states.

* Complete Lecture 61

Added a timer sprite and script.

* Complete Lecture 62

Added a calculation for the fill percentage of the timer image.

* Complete Lecture 63

Connected the timer to the quiz controller. Note that this architecture sucks and I should convert it to an event-driven architecture if I revisit this project in the future.

* Complete Lecture 64

Modified the quiz controller to be able to show multiple questions. Also discovered a neat and slightly annoying quirk of serialized objects.

* Complete Lecture 64

Completely went off the rails to convert the game to using a finite state machine. The tutorial made my head hurt, but everything is back to normal I think 🤞

* Complete Lecture 65

Moved ScoreText handling to QuizController.

* Complete Lecture 66

Added slider to track progress.

* Complete Lecture 67

Added an end screen.

* Complete Lecture 68

Connected the end screen to the rest of the game.